optimization

+++ src/App.js
@@ -9,56 +9,58 @@ import Favorite from "./Pages/Favorite";
 import {logDOM} from "@testing-library/react";
 import AppContext from "./AppContext";
 import OrderPage from "./Pages/OrderPage/OrderPage";
+import favorite from "./Pages/Favorite";
+import order from "./Pages/OrderPage/Order/Order";
 
 function App() {
-    const [cartOpened, setCartOpened] = React.useState(false)
-    const [cartItems, setCartItems] = React.useState([])
-    const [searchInput, setSearchInput] = React.useState('')
-    const [items, setItems] = React.useState([])
     const [favoriteItems, setFavoriteItems] = React.useState([])
+    const [orderItems, setOrderItems] = React.useState([])
+    const [cartItems, setCartItems] = React.useState([])
+    const [items, setItems] = React.useState([])
+    const [searchInput, setSearchInput] = React.useState('')
+    const [orderId, setOrderId] = React.useState(0)
     const [isLoading, setIsLoading] = React.useState(true)
     const [orderIsCreated, setOrderIsCreated] = React.useState(false)
-    const [orderId, setOrderId] = React.useState(1)
-    const [orderIsLoading,setOrderIsLoading]=React.useState(false)
+    const [cartOpened, setCartOpened] = React.useState(false)
+
 
     React.useEffect(() => {
-        const responded = async () => {
+            const responded = async () => {
 
 
-            const itemsRes = await axios.get('https://631a621adc236c0b1edd3f63.mockapi.io/items')
-            const cartRes = await axios.get('https://631a621adc236c0b1edd3f63.mockapi.io/cart')
-            const favoriteRes = await axios.get('https://631a621adc236c0b1edd3f63.mockapi.io/favorite')
+                const itemsRes = await axios.get('https://631a621adc236c0b1edd3f63.mockapi.io/items')
+                const cartRes = await axios.get('https://631a621adc236c0b1edd3f63.mockapi.io/cart')
+                const favoriteRes = await axios.get('https://631a621adc236c0b1edd3f63.mockapi.io/favorite')
+                const orderRes = await axios.get('https://63c1bc2b376b9b2e648305db.mockapi.io/order')
+                setIsLoading(false)
 
-            setIsLoading(false)
-
-            setCartItems(cartRes.data)
-            setFavoriteItems(favoriteRes.data)
-            setItems(itemsRes.data)
-
-
+                setCartItems(cartRes.data)
+                setFavoriteItems(favoriteRes.data)
+                setItems(itemsRes.data)
+                setOrderItems(orderRes.data)
+                orderRes.data.length && setOrderId(Number(orderRes.data[orderRes.data.length - 1].id))
+            }
+            responded()
         }
-        responded()
-    }, [])
+
+        , []
+    )
 
     const onClickCreateOrder = async () => {
         try {
+            if (cartItems.length) {
+                setOrderId(orderItems<1?1:Number(orderItems[orderItems.length-1].id)+1)
+                setOrderIsCreated(true)
 
-            setOrderIsLoading(true)
-            const {data} = await axios.post(`https://63c1bc2b376b9b2e648305db.mockapi.io/order`, {
-                items: [...cartItems]
-            })
-            setCartItems(prevState => [...prevState, data])
-            setCartItems([])
-
-            setOrderId(data.id)
-            setOrderIsCreated(true)
-
-            for (let i = 0; i < cartItems.length; i++) {
-                await axios.delete(`https://631a621adc236c0b1edd3f63.mockapi.io/cart/${cartItems[i].id}`)
-
+                const {data} = await axios.post(`https://63c1bc2b376b9b2e648305db.mockapi.io/order`, {
+                    items: [...cartItems]
+                })
+                setOrderItems(prevState => [...prevState, data])
+                setCartItems([])
+                for (let i = 0; i < cartItems.length; i++) {
+                    await axios.delete(`https://631a621adc236c0b1edd3f63.mockapi.io/cart/${cartItems[i].id}`)
+                }
             }
-            setOrderIsLoading(false)
-
         } catch (error) {
             alert('Не удалось создать заказ')
         }
@@ -69,14 +71,16 @@ function App() {
     }
 
     const onClickAddToCart = async (item) => {
-
         try {
             const foundedItem = cartItems.find((cartItem) => cartItem.name === item.name)
             if (foundedItem) {
-                removeFromDrawerCart(foundedItem.id)
+                axios.delete(`https://631a621adc236c0b1edd3f63.mockapi.io/cart/${foundedItem.id}`)
+                setCartItems(prevState => [...prevState.filter(item => item.id !== foundedItem.id)])
+
             } else {
+                setCartItems(prevState => [...prevState, item])
                 const {data} = await axios.post(`https://631a621adc236c0b1edd3f63.mockapi.io/cart`, item)
-                setCartItems(prevState => [...prevState, data])
+                setCartItems(prevState => [...prevState.slice(0, -1), data])
             }
         } catch (error) {
             alert('Не удалось добавить в корзину')
@@ -86,6 +90,7 @@ function App() {
     const onClickAddToFavorite = async (item, inFavoritePage = false) => {
         try {
             const foundedItem = favoriteItems.find((favoriteItem) => favoriteItem.name === item.name)
+
             if (foundedItem) {
                 axios.delete(`https://631a621adc236c0b1edd3f63.mockapi.io/favorite/${foundedItem.id}`)
                 if (!inFavoritePage) setFavoriteItems(prevState => [...prevState.filter(item => item.name !== foundedItem.name)])
@@ -103,9 +108,18 @@ function App() {
     const itemIsAddedCart = (name) => {
         return cartItems.some((cartItem) => cartItem.name === name)
     }
+    const itemIsAddedFavorite = (name) => {
+        return favoriteItems.some((favoriteItem) => favoriteItem.name === name)
+    }
+
     return (
 
-        <AppContext.Provider value={{cartItems, favoriteItems, itemIsAddedCart,onClickAddToCart}}>
+        <AppContext.Provider value={{
+            cartItems, favoriteItems, itemIsAddedCart,
+            itemIsAddedFavorite, onClickAddToCart, orderIsCreated,
+            setOrderIsCreated, orderItems, setOrderItems,
+            setOrderId
+        }}>
             <div className="wrapper clear">
                 {cartOpened && <DrawerCart onClickDelete={removeFromDrawerCart}
                                            onClickOverlay={() => {
@@ -115,11 +129,11 @@ function App() {
                                            onClickCreateOrder={onClickCreateOrder}
                                            orderIsCreated={orderIsCreated}
                                            orderId={orderId}
-                                           orderIsLoading={orderIsLoading}
                 />}
 
                 <Header onClickCart={() => setCartOpened(true)}
                         cartItems={cartItems}
+                        orderIsCreated={orderIsCreated}
                 />
 
                 <Routes>
